Executed final step of migration to standalone
