+++ src/app/app.component.ts
@@ -1,10 +1,19 @@
 import { Component } from '@angular/core';
 import { GalleryFacade } from './state-management/gallery-list/gallery.facade';
+import { FooterComponent } from './components/footer/footer.component';
+import { RouterOutlet } from '@angular/router';
+import { HeaderComponent } from './components/header/header.component';
 
 @Component({
-  selector: 'app-root',
-  templateUrl: './app.component.html',
-  styleUrls: ['./app.component.scss'],
+    selector: 'app-root',
+    templateUrl: './app.component.html',
+    styleUrls: ['./app.component.scss'],
+    standalone: true,
+    imports: [
+        HeaderComponent,
+        RouterOutlet,
+        FooterComponent,
+    ],
 })
 export class AppComponent {
   constructor(private galleryFacade: GalleryFacade) {
